perf(badges): defer heavy mount past the entry transition + skeleton
Opening the Badges tab froze the source screen: it mounted useDrinkEvents (the
dominant buildDrinkEvents pass) synchronously during the navigation slide.
Adopt the StatisticsScreen / DayOverview pattern — gate the heavy body on the
screen's transitionEnd via useReadyAfterScreenTransition +
ScreenWrapper.onEntryTransitionEnd (runAfterInteractions resolves before the
slide finishes, so it can't fix this).

- BadgesScreen is now a thin parent that shows BadgesScreenSkeleton until the
  slide ends, then mounts BadgesContent (useDrinkEvents + streak/stat/badge UI).
- BadgesScreenSkeleton is a layout-faithful placeholder on the shared shimmering
  Skeleton primitive (ChartSkeleton kpi + kpiRow + badge rows).
